+++ core/src/main/java/org/lemurproject/galago/core/index/mem/MemoryCountIndex.java
@@ -458,7 +458,7 @@ public class MemoryCountIndex implements MemoryIndexPart, AggregateIndexPart {
     }
 
     @Override
-    public void syncTo(int identifier) throws IOException {
+    public void syncTo(long identifier) throws IOException {
       // TODO: need to implement skip lists
 
       while (!isDone() && (currDocument < identifier)) {
@@ -467,7 +467,7 @@ public class MemoryCountIndex implements MemoryIndexPart, AggregateIndexPart {
     }
 
     @Override
-    public void movePast(int identifier) throws IOException {
+    public void movePast(long identifier) throws IOException {
       while (!isDone() && (currDocument <= identifier)) {
         read();
       }
