+++ core/src/main/java/org/lemurproject/galago/core/index/mem/MemoryDocumentLengths.java
@@ -366,12 +366,12 @@ public class MemoryDocumentLengths implements MemoryIndexPart, LengthsReader {
     }
 
     @Override
-    public void movePast(int identifier) throws IOException {
+    public void movePast(long identifier) throws IOException {
       syncTo(identifier + 1);
     }
 
     @Override
-    public void syncTo(int identifier) throws IOException {
+    public void syncTo(long identifier) throws IOException {
       this.currDoc = identifier;
       if (identifier > this.fieldLengths.lastDocument) {
         done = true;
